% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_data.R
\docType{data}
\name{Integration_Tables}
\alias{Integration_Tables}
\title{Tables of integrations used for fast computation}
\usage{
data(Integration_Tables)
}
\description{
A dataset containing one normal and two logistic integration tables
}
\details{
#'@format A list containing tables of integrations for normal, logistic and Cauchy distributions:
\describe{
  \item{normal.table}{769-by-3 table for a normal distribution}
  \item{logistic.table1}{2561-by-3 table for the logistic distribution}
  \item{logistic.table2}{3841-by-1 table for the logistic distribution}
  \item{cauchy.table}{2561-by-3 table for the Cauchy distribution}
}
}
\keyword{datasets}
